change int to long for manufacturerid

+++ src/main/java/org/jsack/tifsa/Database/Manufacturer/Manufacturer.java
@@ -4,11 +4,11 @@ package org.jsack.tifsa.Database.Manufacturer;
  * Created by julius on 3/21/2017.
  */
 public class Manufacturer {
-    private int manufacturerID;
+    private long manufacturerID;
     private String manufacturerName;
 
-    public int getManufacturerID(){return manufacturerID;}
-    public void setManufacturerID(int manufacturerID){this.manufacturerID = manufacturerID;}
+    public long getManufacturerID(){return manufacturerID;}
+    public void setManufacturerID(long manufacturerID){this.manufacturerID = manufacturerID;}
 
     public String getManufacturerName(){return manufacturerName;}
     public void setManufacturerName(String manufacturerName){this.manufacturerName = manufacturerName;}
